cortex/requirements.txt
attrdict==2.0.1
attrs==19.3.0
base58==2.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cytoolz==0.10.1
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.3
Flask==1.0.3
Flask-Cors==3.0.8
hexbytes==0.2.1
idna==2.8
importlib-metadata==1.7.0
ipfshttpclient==0.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
pkg-resources==0.0.0
protobuf==3.13.0
pycryptodome==3.9.8
pyrsistent==0.16.0
python-dotenv==0.11.0
requests==2.22.0
rlp==1.2.0
six==1.15.0
toolz==0.10.0
urllib3==1.25.10
varint==1.0.2
web3==5.2.0
websockets==7.0
Werkzeug==0.16.0
zipp==3.1.0
